requirements_dev.txt
Cython>=0.24.1
numpy>=1.19.0
pandas>=1.1.2
scikit-learn>=0.22
scikit-surprise>=1.1.1
joblib>=0.17
six>=1.15.0